move progress into a core package
