+++ alive_progress/core/progress.py
@@ -9,8 +9,8 @@ from contextlib import contextmanager
 from datetime import timedelta
 from itertools import chain, islice, repeat
 
-from .animations.utils import spinner_player
-from .configuration import config_handler
+from ..animations.utils import spinner_player
+from ..configuration import config_handler
 
 
 @contextmanager
